+++ blackdoc/formats/doctest.py
@@ -1,6 +1,8 @@
+import ast
 import io
 import itertools
 import re
+import sys
 import tokenize
 from tokenize import TokenError
 
@@ -174,6 +176,42 @@ def restore_quotes(code_unit, original_quotes):
     return restored_code_unit
 
 
+def split_by_statement(code_unit):
+    """split a code unit into individual statements
+
+    At this point, the only way to have more than a single statement
+    is by joining multiple (non-block) statements with a `;`.
+    """
+
+    def lineno(node):
+        # TODO: remove once we drop support for python=3.7
+        version = (sys.version_info.major, sys.version_info.minor)
+
+        if (
+            version < (3, 8)
+            and isinstance(node, ast.Expr)
+            and isinstance(node.value, ast.Str)
+        ):
+            # bug in ast (fixed in py38): lineno is wrong for multi-line string expressions
+            # https://bugs.python.org/issue16806
+            n_lines = len(node.value.s.split("\n"))
+            lineno = node.lineno - n_lines + 1
+        else:
+            lineno = node.lineno
+
+        return lineno
+
+    content = ast.parse(code_unit).body
+
+    lines = code_unit.split("\n")
+    if not content:
+        return [lines]
+
+    indices = [lineno(obj) - 1 for obj in content]
+    slices = more_itertools.zip_offset(indices, indices, offsets=(0, 1), longest=True)
+    return [lines[start:stop] for start, stop in slices]
+
+
 def reformatting_func(code_unit, docstring_quotes):
     def add_prompt(prompt, line):
         if not line:
@@ -181,20 +219,22 @@ def reformatting_func(code_unit, docstring_quotes):
 
         return " ".join([prompt, line])
 
-    restored_quotes = restore_quotes(code_unit, docstring_quotes)
+    def reformat_code_unit(lines):
+        if block_start_re.match(lines[0]):
+            lines.append("")
 
-    lines = restored_quotes.rstrip().split("\n")
-    if block_start_re.match(lines[0]):
-        lines.append("")
-
-    lines_ = iter(lines)
-    reformatted = list(
-        itertools.chain(
+        lines_ = iter(lines)
+        return itertools.chain(
             more_itertools.always_iterable(
                 add_prompt(prompt, more_itertools.first(lines_))
             ),
             (add_prompt(continuation_prompt, line) for line in lines_),
         )
-    )
 
-    return "\n".join(reformatted)
+    restored_quotes = restore_quotes(code_unit.rstrip(), docstring_quotes)
+
+    subunits = split_by_statement(restored_quotes)
+
+    return "\n".join(
+        itertools.chain.from_iterable(reformat_code_unit(unit) for unit in subunits)
+    )